use bigint instead of regular 32-bit int

+++ mlos_bench/mlos_bench/storage/sql_schema.py
@@ -9,7 +9,7 @@ DB schema definition.
 import logging
 
 from sqlalchemy import (
-    MetaData, Table, Column, Sequence, Integer, String, DateTime,
+    MetaData, Table, Column, Sequence, BigInteger, String, DateTime,
     PrimaryKeyConstraint, ForeignKeyConstraint, UniqueConstraint,
 )
 
@@ -43,11 +43,11 @@ class DbSchema:
         # A workaround to support autoincrement in DuckDB:
         if engine.dialect.name == "duckdb":
             seq_config_id = Sequence('seq_config_id')
-            col_config_id = Column("config_id", Integer, seq_config_id,
+            col_config_id = Column("config_id", BigInteger, seq_config_id,
                                 server_default=seq_config_id.next_value(),
                                 nullable=False, primary_key=True)
         else:
-            col_config_id = Column("config_id", Integer, nullable=False,
+            col_config_id = Column("config_id", BigInteger, nullable=False,
                                    primary_key=True, autoincrement=True)
 
         self.config = Table(
@@ -61,8 +61,8 @@ class DbSchema:
             "trial",
             self.meta,
             Column("exp_id", String(255), nullable=False),
-            Column("trial_id", Integer, nullable=False),
-            Column("config_id", Integer, nullable=False),
+            Column("trial_id", BigInteger, nullable=False),
+            Column("config_id", BigInteger, nullable=False),
             Column("ts_start", DateTime, nullable=False, default="now"),
             Column("ts_end", DateTime),
             # Should match the text IDs of `mlos_bench.environment.Status` enum:
@@ -78,7 +78,7 @@ class DbSchema:
         self.config_param = Table(
             "config_param",
             self.meta,
-            Column("config_id", Integer, nullable=False),
+            Column("config_id", BigInteger, nullable=False),
             Column("param_id", String(255), nullable=False),
             Column("param_value", String(255)),
 
@@ -92,7 +92,7 @@ class DbSchema:
             "trial_param",
             self.meta,
             Column("exp_id", String(255), nullable=False),
-            Column("trial_id", Integer, nullable=False),
+            Column("trial_id", BigInteger, nullable=False),
             Column("param_id", String(255), nullable=False),
             Column("param_value", String(255)),
 
@@ -105,7 +105,7 @@ class DbSchema:
             "trial_result",
             self.meta,
             Column("exp_id", String(255), nullable=False),
-            Column("trial_id", Integer, nullable=False),
+            Column("trial_id", BigInteger, nullable=False),
             Column("metric_id", String(255), nullable=False),
             Column("metric_value", String(255)),
 
@@ -118,7 +118,7 @@ class DbSchema:
             "trial_telemetry",
             self.meta,
             Column("exp_id", String(255), nullable=False),
-            Column("trial_id", Integer, nullable=False),
+            Column("trial_id", BigInteger, nullable=False),
             Column("ts", DateTime, nullable=False, default="now"),
             Column("metric_id", String(255), nullable=False),
             Column("metric_value", String(255)),
